tests/wpt/web-platform-tests/css/css-images/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of CSS Images properties</title>
<link rel="help" href="https://drafts.csswg.org/css-images-3/#property-index">
<meta name="assert" content="Properties inherit or not according to the spec.">
<meta name="assert" content="Properties have initial values according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
<div id="target"></div>
</div>
<script>
assert_inherited('image-orientation', '0deg', 'from-image'); // none is 0deg
assert_inherited('image-rendering', 'auto', 'pixelated');
assert_not_inherited('object-fit', 'fill', 'contain');
assert_not_inherited('object-position', '50% 50%', '10px 20px');
</script>
</body>
</html>
